refactor: move card and dialogs in a separate package
