+++ src/main/java/edu/ntnu/idatt2003/millions/view/dashboard/watchlist/WatchlistView.java
@@ -4,6 +4,7 @@ import edu.ntnu.idatt2003.millions.controller.TradeController;
 import edu.ntnu.idatt2003.millions.service.GameService;
 import edu.ntnu.idatt2003.millions.keyboard.SearchFocusProvider;
 import edu.ntnu.idatt2003.millions.view.component.ExploreStocksButton;
+import edu.ntnu.idatt2003.millions.view.dashboard.watchlist.card.WatchlistCard;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
